+++ src/SSB_RetuningAutomations.py
@@ -1375,6 +1375,8 @@ def main():
 
     print("\nLoading Tool...")
     # Remove Splash image from Pyinstaller
+    import os
+    import importlib
     if '_PYI_SPLASH_IPC' in os.environ and importlib.util.find_spec("pyi_splash"):
         import pyi_splash
         pyi_splash.update_text('UI Loaded ...')